최규원 방삭제 구현

+++ src/main/java/com/project/roomescape/responseDto/RoomResponseDto.java
@@ -10,9 +10,9 @@ public class RoomResponseDto {
     private String teamName;
     private Long count;
     private String createdUser;
-    private Long currentNum;    // 현재 참여 인원
+    private Integer currentNum;    // 현재 참여 인원
 
-    public RoomResponseDto(Long roomId, String teamName, Long count, String createdUser, Long currentNum) {
+    public RoomResponseDto(Long roomId, String teamName, Long count, String createdUser, Integer currentNum) {
         this.roomId = roomId;
         this.teamName = teamName;
         this.count = count;
